chore(uagents): move quota protocol to main protocol module
